chore: moving minigames to ts
