+++ src/games/base-scenes/scene-in-metagym-room.ts
@@ -1,6 +1,6 @@
 import { EarnableScene } from "./EarnableScene";
 import Phaser from "phaser";
-import { GYM_ROOM_SCENE } from "../../shared";
+import { GYM_ROOM_SCENE } from "../shared";
 
 type handleExitParams = {
   thisSceneKey: string;
