+++ src/games/minigames/MatrixScene.ts
@@ -5,7 +5,7 @@ import { MATRIX } from "../shared";
 import { FONT, PILL_BLUE, PILL_RED } from "../gym-room-boot/assets";
 import { createTextBox } from "../utils/text";
 import { mainBgColorNum, highlightTextColorNum } from "../../GlobalStyles";
-import { SceneInMetaGymRoom } from "../base-scenes/ts/scene-in-metagym-room";
+import { SceneInMetaGymRoom } from "../base-scenes/scene-in-metagym-room";
 
 const SceneConfig = {
   active: false,
