+++ src/games/minigames/SpaceStretchScene.ts
@@ -13,7 +13,7 @@ import {
   InGameFont,
 } from "../../GlobalStyles";
 import party, { sources } from "party-js";
-import { SceneInMetaGymRoom } from "../base-scenes/ts/scene-in-metagym-room";
+import { SceneInMetaGymRoom } from "../base-scenes/scene-in-metagym-room";
 
 const SceneConfig = {
   active: false,
